Move test logger to own module
