+++ e2e/nx-firebase-e2e/test-utils/index.ts
@@ -2,8 +2,10 @@
 import { joinPathFragments, names } from '@nx/devkit'
 import { readJson, runNxCommandAsync } from '@nx/plugin/testing'
 import { expectStrings, expectNoStrings } from './test-helpers'
+import { testDebug, red, green } from './test-logger'
 
 const NPM_SCOPE = '@proj'
+const STRIP_ANSI_MATCHER = /[\u001b\u009b][[()#;?]*(?:[0-9]{1,4}(?:;[0-9]{0,4})*)?[0-9A-ORZcf-nqry=><]/g
 
 export interface ProjectData {
   name: string
@@ -17,26 +19,6 @@ export interface ProjectData {
   configName: string
 }
 
-const ENABLE_TEST_DEBUG_INFO = true
-const STRIP_ANSI_MATCHER = /[\u001b\u009b][[()#;?]*(?:[0-9]{1,4}(?:;[0-9]{0,4})*)?[0-9A-ORZcf-nqry=><]/g
-
-// https://stackoverflow.com/questions/9781218/how-to-change-node-jss-console-font-color
-const GREEN_FG = '\x1b[32m'
-const RED_FG = '\x1b[31m'
-const RESET_FG = '\x1b[0m'
-
-export function green(text: string) {
-  return `${GREEN_FG}${text}${RESET_FG}`
-}
-export function red(text: string) {
-  return `${RED_FG}${text}${RESET_FG}`
-}
-
-export function testDebug(info: string) {
-  if (ENABLE_TEST_DEBUG_INFO) {
-    console.debug(info)
-  }
-}
 
 export async function safeRunNxCommandAsync(cmd: string)
 {
